Fixed size limit, tests
