+++ create-i18n/index.test.ts
@@ -236,237 +236,4 @@ test('tracks double definition', () => {
   }, /I18n component double was defined multiple times/)
 })
 
-/* 
------------ Partial loading testing -----------
-*/
-
-let partialGetCalls: object[] = []
-let partialResolveGet: (translations: ComponentsJSON[]) => void = () => {}
-let partialRequestsByLocale: Record<string, typeof partialResolveGet> = {}
-
-function partialGet(
-  code: string,
-  components: string[]
-): Promise<ComponentsJSON[]> {
-  partialGetCalls.push({ [code]: components })
-  return new Promise(resolve => {
-    partialRequestsByLocale[code] = resolve
-    partialResolveGet = resolve
-  })
-}
-
-async function partialGetResponse(
-  translations: ComponentsJSON[],
-  code?: string
-): Promise<void> {
-  if (code) {
-    partialRequestsByLocale[code](translations)
-  } else {
-    partialResolveGet(translations)
-  }
-}
-
-test.after.each(() => {
-  partialGetCalls = []
-})
-
-let locale = atom<'en' | 'ru' | 'fr' | 'de'>('ru')
-let i18n = createI18n(locale, { get: partialGet })
-let events: string[] = []
-
-let post = i18n('main/post', { title: 'Post' })
-let heading = i18n('main/heading', { title: 'Title' })
-let comment = i18n('main/comment', { title: 'Comment' })
-let message = i18n('chat/message', { title: 'Message' })
-let user = i18n('settings/user', { title: 'User' })
-let games = i18n('games', { title: 'Games' })
-
-let postUnbind: () => void
-let headingUnbind: () => void
-let commentUnbind: () => void
-
-test("after components mount shouldn't load translations with same prefix in componentName", async () => {
-  equal(i18n.loading.get(), false)
-  equal(partialGetCalls, [])
-
-  postUnbind = post.subscribe(t => {
-    events.push(t.title)
-  })
-  headingUnbind = heading.subscribe(t => {
-    events.push(t.title)
-  })
-  equal(i18n.loading.get(), true)
-  equal(partialGetCalls, [{ ru: ['main/post'] }])
-  equal(events, ['Post', 'Title'])
-})
-
-test('loads translations partial', async () => {
-  await partialGetResponse([
-    {
-      'main/post': { title: 'Публикация' },
-      'main/heading': { title: 'Заголовок' },
-      'main/comment': { title: 'Комментарий' }
-    }
-  ])
-  equal(i18n.loading.get(), false)
-  equal(events, ['Post', 'Title', 'Публикация', 'Заголовок'])
-  equal(post.get(), { title: 'Публикация' })
-  equal(heading.get(), { title: 'Заголовок' })
-  equal(comment.get(), { title: 'Комментарий' })
-  events = []
-})
-
-test('after components mount should load translations with different prefix or without prefix names', async () => {
-  equal(i18n.loading.get(), false)
-  equal(partialGetCalls, [])
-
-  let userUnbind = user.subscribe(t => {
-    events.push(t.title)
-  })
-  let gamesUnbind = games.subscribe(t => {
-    events.push(t.title)
-  })
-  equal(i18n.loading.get(), true)
-  equal(partialGetCalls, [{ ru: ['settings/user'] }, { ru: ['games'] }])
-  equal(events, ['User', 'Games'])
-
-  await partialGetResponse([
-    {
-      'settings/user': { title: 'Пользователь' },
-      'games': { title: 'Игры' }
-    }
-  ])
-  equal(i18n.loading.get(), false)
-  equal(events, [
-    'User',
-    'Games',
-    'Публикация',
-    'Заголовок',
-    'Пользователь',
-    'Игры'
-  ])
-  userUnbind()
-  gamesUnbind()
-  events = []
-})
-
-test("component mounting shouldn't load cached translations", async () => {
-  commentUnbind = comment.subscribe(t => {
-    events.push(t.title)
-  })
-  equal(i18n.loading.get(), false)
-  equal(partialGetCalls, [])
-  equal(events, ['Комментарий'])
-  events = []
-})
-
-test('loads translation after component mounted', async () => {
-  message.subscribe(t => {
-    events.push(t.title)
-  })
-  equal(i18n.loading.get(), true)
-  equal(partialGetCalls, [{ ru: ['chat/message'] }])
-  equal(events, ['Message'])
-
-  await partialGetResponse([
-    {
-      'chat/message': { title: 'Сообщение' }
-    }
-  ])
-  equal(i18n.loading.get(), false)
-  equal(events, [
-    'Message',
-    'Публикация',
-    'Заголовок',
-    'Комментарий',
-    'Сообщение'
-  ])
-  equal(message.get(), { title: 'Сообщение' })
-  events = []
-})
-
-test("locale changing shouldn't load cached translations", async () => {
-  locale.set('en')
-  equal(i18n.loading.get(), false)
-  equal(partialGetCalls, [])
-  equal(events, ['Post', 'Title', 'Comment', 'Message'])
-  events = []
-})
-
-test('locale changing should load translations for mounted components only', async () => {
-  postUnbind()
-  headingUnbind()
-  commentUnbind()
-
-  // 1s delay for components unmounting
-  await new Promise(resolve => setTimeout(resolve, 1100))
-
-  locale.set('fr')
-  equal(i18n.loading.get(), true)
-  equal(partialGetCalls, [{ fr: ['chat/message'] }])
-  equal(events, [])
-
-  await partialGetResponse([
-    {
-      'chat/message': { title: 'Le message' }
-    }
-  ])
-  equal(i18n.loading.get(), false)
-  equal(events, ['Le message'])
-  equal(post.get(), { title: 'Post' })
-  equal(heading.get(), { title: 'Title' })
-  equal(comment.get(), { title: 'Comment' })
-  equal(message.get(), { title: 'Le message' })
-  events = []
-})
-
-test('if get returns array of ComponentsJSON, it transforms to object before cached', async () => {
-  locale.set('en')
-
-  post.subscribe(t => {
-    events.push(t.title)
-  })
-  heading.subscribe(t => {
-    events.push(t.title)
-  })
-  comment.subscribe(t => {
-    events.push(t.title)
-  })
-
-  locale.set('de')
-  equal(i18n.loading.get(), true)
-  equal(partialGetCalls, [
-    { de: ['chat/message', 'main/post', 'main/heading', 'main/comment'] }
-  ])
-  equal(events, ['Message', 'Post', 'Title', 'Comment'])
-
-  await partialGetResponse([
-    {
-      'main/post': { title: 'Publikation' },
-      'main/heading': { title: 'Titel' },
-      'main/comment': { title: 'Kommentar' }
-    },
-    {
-      'chat/message': { title: 'Nachricht' }
-    }
-  ])
-  equal(i18n.cache.de, {
-    'main/post': { title: 'Publikation' },
-    'main/heading': { title: 'Titel' },
-    'main/comment': { title: 'Kommentar' },
-    'chat/message': { title: 'Nachricht' }
-  })
-  equal(i18n.loading.get(), false)
-  equal(events, [
-    'Message',
-    'Post',
-    'Title',
-    'Comment',
-    'Nachricht',
-    'Publikation',
-    'Titel',
-    'Kommentar'
-  ])
-})
-
 test.run()
